+++ src/Math/Vec3.cpp
@@ -10,11 +10,11 @@ Vec3 Vec3::operator-(const Vec3& v) const
 {
 	return { x() - v.x(), y() - v.y(), z() - v.z() };
 }
-double Vec3::operator*(const Vec3& v) const
+float Vec3::operator*(const Vec3& v) const
 {
 	return x() * v.x() + y() * v.y() + z() * v.z();
 }
-Vec3 Vec3::operator/(const double v) const
+Vec3 Vec3::operator/(const float v) const
 {
 	return { x() / v, y() / v, z() / v };
 }
@@ -33,14 +33,14 @@ Vec3& Vec3::operator-=(const Vec3& v)
 	this->val[2] -= v[2];
 	return *this;
 }
-Vec3& Vec3::operator*=(const double f)
+Vec3& Vec3::operator*=(const float f)
 {
 	this->val[0] *= f;
 	this->val[1] *= f;
 	this->val[2] *= f;
 	return *this;
 }
-Vec3& Vec3::operator/=(const double f)
+Vec3& Vec3::operator/=(const float f)
 {
 	this->val[0] /= f;
 	this->val[1] /= f;
@@ -57,7 +57,7 @@ Vec3 Vec3::cross(const Vec3& v) const
 	};
 }
 
-double Vec3::length() const
+float Vec3::length() const
 {
 	return std::sqrt(x() * x() + y() * y() + z() * z());
 }
@@ -68,20 +68,20 @@ Vec3 Vec3::normalized() const
 	return { x() / l, y() / l, z() / l };
 }
 
-double Vec3::det(const Vec3& v1, const Vec3& v2, const Vec3& v3)
+float Vec3::det(const Vec3& v1, const Vec3& v2, const Vec3& v3)
 {
-	const double x = v2[1] * v3[2] - v3[1] * v2[2];
-	const double y = v2[0] * v3[2] - v3[0] * v2[2];
-	const double z = v2[0] * v3[1] - v3[0] * v2[1];
+	const float x = v2[1] * v3[2] - v3[1] * v2[2];
+	const float y = v2[0] * v3[2] - v3[0] * v2[2];
+	const float z = v2[0] * v3[1] - v3[0] * v2[1];
 
 	return v1[0] * x - v1[1] * y + v1[2] * z;
 }
 
-Vec3 operator*(const double x, const Vec3& v)
+Vec3 operator*(const float x, const Vec3& v)
 {
 	return { x * v.x(), x * v.y(), x * v.z() };
 }
-Vec3 operator*(const Vec3& v, const double x)
+Vec3 operator*(const Vec3& v, const float x)
 {
 	return { x * v.x(), x * v.y(), x * v.z() };
 }
